pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["DarthZiel <Studerzhan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.4"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
drf-yasg = "^1.21.8"
djoser = "^2.3.1"
django-cors-headers = "^4.6.0"
groq = "^0.13.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
